[Entity] Move Entity classes outside pycore/
