+++ nixio/tag.py
@@ -11,7 +11,7 @@ import numpy as np
 
 from .value import DataType
 from .data_array import DataArray
-from .pycore.entity_with_sources import EntityWithSources
+from .entity_with_sources import EntityWithSources
 from .data_view import DataView
 from .pycore.feature import Feature
 from .pycore.exceptions import (OutOfBounds, IncompatibleDimensions,
